Update x_recorder.py
With these changes, the script will now save the downloaded files to a subfolder within the default download directory (/mnt/e/AV/Capture/X-Recorder/). The subfolder will be named after the space ID.
For example, if the space URL is https://x.com/i/spaces/1gqxvNdvyVzxB, the downloaded files will be saved in /mnt/e/AV/Capture/X-Recorder/1gqxvNdvyVzxB/.
The default download directory is now defined as a constant (DEFAULT_DOWNLOAD_DIR) at the beginning of the script, making it easier to modify if needed.